Exit or error with error code
Currently, if an important step in the setup process fails, it is logged
by the setup process continues.  This means that we are having to look
at the logs after each build (for each target) manually to find out if
a particular step as failed.  It is better for the build to fail than a
small step failing silently and causing an incorrect build that is hard
to detect.

Doing this is not likely to cause build failures right away because we
have been monitoring the build logs and didn't see any failure during
last release.  Also, the number of scripts present are quite small now.