Remove first-run completely and avoid rebooting
There are no longer any first-run steps with all the steps having moved to
Plinth. In future, we want to run setup steps such as these in Plinth instead.
Rebooting after running the first-run scripts currently causes problems for
Plinth as it does its setup in background, the setup process is interrupted.

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>
Signed-off-by: Joseph Nuthalapati <njoseph@thoughtworks.com>